Add PrimitiveRepr trait.
This PR adds an additional trait, and a derive macro for automatically
implementing a `PrimitiveRepr` trait with an associated type `Type =T`
that represents the contents of repr attribute: `#[repr(T)]`.

This is useful in certain scenarios where you want to access the
discriminant value in a safe manner, and also ensure that type matches
the contents of `repr(T)` attribute for safety.